Using javascript map instead.
